pyproject.toml
[build-system]
requires = ["pdm-backend<3"]
build-backend = "pdm.backend"

[project]
name = "china-beancount-importers"
version = "0.0.5"
description = "A set of importers maybe useful for you in China."
authors = [{ name = "Trim21", email = "trim21.me@gmail.com" }]
license = "MIT"
requires-python = '>=3.10,<4'
dependencies = [
  "beancount >=3.0.0,<4",
  "beangulp>=0.2.0",
  "beautifulsoup4>=4.12.0",
  "pandas[excel]",
  "pdfplumber>=0.11.0",
  "pydantic>=2.6.0",
  "python-dateutil>=2.9.0",
  "rjieba<1",
  "sslog>=0.0.0a1",
]

[project.urls]
repository = 'https://github.com/trim21/china-beancount-importers'

[dependency-groups]
docs = [
  'sphinx==8.2.3; python_version>="3.12"',
  'sphinx-rtd-theme==3.0.2; python_version>="3.12"',
  'recommonmark==0.7.1 ; python_version>="3.12"',
]
dev = [
  # tests
  "coverage ==7.8.0",
  "pytest>=8.3.5",
  "pytest-github-actions-annotate-failures>=0.3.0",
  'pandas-stubs==3.0.0.260204 ; python_version>="3.11"',
  'types-python-dateutil',
  # lint
  "mypy>=1.15.0",
]

[tool.pytest.ini_options]
addopts = '-rav'

[tool.ruff.lint]
extend-select = ['I']

[tool.uv]
required-environments = [
  "implementation_name == 'cpython' and sys_platform == 'darwin' and platform_machine == 'x86_64'",
  """
  implementation_name == 'cpython'
  and (platform_machine == 'x86_64' or platform_machine == 'AMD64')
  and sys_platform == 'win32'
  """,
  """
  implementation_name == 'cpython'
  and (platform_machine == 'x86_64' or platform_machine == 'AMD64' or platform_machine == 'aarch64')
  and sys_platform == 'linux'
  """,
]
